platformsvc: extract shared serial stream; slim btserial.go to BT surface
